refactor(utils): move ReadOnly out of testonly.go
Production types embed ReadOnly for TestEqual; keep the marker (and
isReadOnly) in a non-testonly file.
